requirements.txt
python-dotenv==1.0.0
openai==1.0.0
google-api-python-client==2.0.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.0
pyyaml==6.0.1
jinja2==3.1.2
pymongo==4.6.1
cohere==5.15.0